Add last update timestamp to UserData, Settings, and Subjects
This commit introduces a `lastUpdate` timestamp field to `NetworkSubjects` and `Settings` models in the `sync_domain`. This field, of type `com.google.firebase.Timestamp`, will be populated with `Timestamp.now()` during uploads in `FirebaseSyncRepository` and `PushSettings`.

Additionally, `FirebaseUserData` now includes `createdAt` and `updatedAt` timestamps.
- `createdAt` is set when new user data is created.
- `updatedAt` is updated using `FieldValue.serverTimestamp()` whenever any user data field is modified. A new private helper function `updateField` was introduced in `FirebaseUserDataRepository` to encapsulate this logic.

Finally, the `firebase-firestore` dependency was added to the `sync_domain` module.