Pass Activity to signIn function and update credentials version
This commit updates the `signIn` function in `GoogleAuthUiClient`, `AuthService`, and `GoogleAuthService` to accept an `Activity` as a parameter. This is necessary because the `credentialManager.getCredential` function now requires an `Activity` context.

The `OnBoardingScreen` has been updated to pass the current `Activity` when calling `authService.signIn()`.

Additionally, the `credentials` library version has been updated from `1.6.0-alpha04` to `1.6.0-alpha05` in `gradle/libs.versions.toml`.