Refactor navigation from AppOnBoarding to Home
This commit refactors the navigation logic when transitioning from the `AppOnBoardingGPASubjectsSettings` screen to the main application graph.

Instead of popping up to a specific destination within the onboarding graph, the code now repeatedly calls `popBackStack()` until the back stack is clear, and then navigates to the start destination of the root navigation graph. This ensures a cleaner navigation stack when entering the main part of the app.